Refactored the main loop and rendering of playin area
